+++ RELEASENOTES.md
@@ -23,6 +23,8 @@
     *   Fix `NullPointerException` in `StyledPlayerView` that could occur after
         calling `StyledPlayerView.setPlayer(null)`
         ([#8985](https://github.com/google/ExoPlayer/issues/8985)).
+*   RTSP:
+    *   Add support for RTSP basic and digest authentication.
 
 ### 2.14.0 (2021-05-13)